Fix admin poll close button
Fixed the poll close button on the admin page. The data being passed back to the server
upon closing the poll was hard coded, now it dynamically adapts to the poll's name and sends that information
bac to the server to close the poll and issue the sockets broadcast to all clients to disable voting.